Move bearer token method out of TestAccounts.
